VBA first project
I ended up learning a lot with this project, it took me 3 days to learn the basics of VBA and put all of this together. (Note, I was learning java at the time, This was probably 6 months after I started to code outside Matlab which was how i was taught to code back in college.